small changes in response to tridge's comments, including moving soaring_controller back to the G2 parameter group
