+++ ArduPlane/Parameters.cpp
@@ -1167,11 +1167,7 @@ const AP_Param::Info Plane::var_info[] = {
     // @Group: NTF_
     // @Path: ../libraries/AP_Notify/AP_Notify.cpp
     GOBJECT(notify, "NTF_",  AP_Notify),
-  
-    // @Group: SOAR_
-    // @Path: ../libraries/AP_Soaring/AP_Soaring.cpp
-    GOBJECT(soaring_controller, "SOAR_", SoaringController),
-  
+
     // @Group: 
     // @Path: Parameters.cpp
     GOBJECT(g2, "",  ParametersG2),
@@ -1216,12 +1212,17 @@ const AP_Param::GroupInfo ParametersG2::var_info[] = {
     // @Group: RC
     // @Path: ../libraries/RC_Channel/RC_Channels.cpp
     AP_SUBGROUPINFO(rc_channels, "RC", 7, ParametersG2, RC_Channels),
+    
+    // @Group: SOAR_
+    // @Path: ../libraries/AP_Soaring/AP_Soaring.cpp
+    AP_SUBGROUPINFO(soaring_controller, "SOAR_", 8, ParametersG2, SoaringController),
   
     AP_GROUPEND
 };
 
 ParametersG2::ParametersG2(void) :
-    ice_control(plane.rpm_sensor, plane.ahrs) 
+    ice_control(plane.rpm_sensor, plane.ahrs),
+    soaring_controller(plane.ahrs, plane.SpdHgt_Controller, plane.aparm)
 {
     AP_Param::setup_object_defaults(this, var_info);
 }
